Fix FileLockProvider, deleting files defeats its purpose
Re-implement `NioFileLockProvider` in order to use a single file to hold
all the locks, using region locking on a stripped bucket list.

The approach of using a single file per lock requires deleting the files
at some point to avoid flooding the locks dir, but there's no way around
in breaking distributed locking that way. That is, when a process
deletes one of the files during release(), other processes may still
have an open FileChannel, which won't dissappear until closing the
channel, it will be open against a file that has no name. So when yet
another process acquires a lock, it re-creates the file, and both end up
having a FileLock on the same key, one with the newly created version of
the file, and the other one with the deleted but not yet closed file
channel.